Refactor Lift subsystem: fix bugs and improve type handling
Changed setLiftPosition() parameter from int to double for better precision. Removed unnecessary int casts in Lower and Raise actions. Fixed bug where rightMotor.setPower(1) was called twice instead of setting leftMotor power. Made motors public in LiftIO. Removed unused imports (IntakeIO, DcMotorSimple). Reorganized motor initialization to defer RUN_TO_POSITION mode setting until setLiftPosition() is called. Added proper mode switching in setLiftPwr() and resetEncoders() methods.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Lift/Actions/Lower.java
@@ -22,7 +22,7 @@ public class Lower implements Action {
     public boolean run(@NonNull TelemetryPacket packet) {
 
         if (!initialized) {
-            liftIO.setLiftPosition((int)Constants.LIFT_DOWN);
+            liftIO.setLiftPosition(Constants.LIFT_DOWN);
             initialized = true;
 
         }
